fix: toast message
