+++ pages/mywallet.js
@@ -2,7 +2,7 @@ import Layout from "../component/Layout";
 import Typography from "../component/Typography";
 import NFT from "../component/NFT";
 import axios from "axios";
-import React, { Component, useEffect, useState } from "react";
+import React, { Component, useEffect, useState, useRef } from "react";
 import styled, { css, keyframes } from "styled-components";
 import _ from "lodash";
 import Margin from "../component/Margin";
@@ -11,58 +11,89 @@ import Footer from "../component/Footer";
 import Carousel from "nuka-carousel";
 import Link from "next/link";
 import { useRouter } from "next/router";
+import domtoimage from "dom-to-image";
+import Toast from "../component/Toast";
 
 const BoldTypography = styled(Typography)`
-    font-weight: bold;
+  font-weight: bold;
 `;
 
 const Header = styled.div`
-    width: 300px;
-    text-align: left;
+  width: 300px;
+  text-align: left;
 `;
 
 const NTFborder = styled.div`
-    flex-direction: column;
-    justify-content: center;
-    //box-shadow: -10px 20px 20px #000; // 그림자 이상해서 일단 주석 처리
+  flex-direction: column;
+  justify-content: center;
+  //box-shadow: -10px 20px 20px #000; // 그림자 이상해서 일단 주석 처리
 `;
 
 const Menu = {
-    alignItems: "left",
+  alignItems: "left",
 };
 
+const ShareLogo = styled.img`
+  height: 18px;
+`;
+
+const ShareLogoBox = styled.div`
+  margin-top: 10px;
+  margin-left: 200px;
+  float: right;
+  width: 30px;
+  height: 30px;
+  z-index: 5;
+  position: absolute;
+`;
+
 export default function Wallet() {
-    const [nftData, setNftData] = useState([]);
-    const [cardList, setCardList] = useState();
-    const router = useRouter();
+  const [nftData, setNftData] = useState([]);
+  const [cardList, setCardList] = useState();
+  const [active, setActive] = useState(false); // 저장 완료 토스트 메시지용 State
+  const router = useRouter();
 
-    const [token, setToken] = useState();
+  const [token, setToken] = useState();
 
-    useEffect(() => {
-        if (typeof window !== "undefined") {
-            const item = localStorage.getItem("token");
-            setToken(item);
-            if (!item) {
-                router.push("/login");
-            }
-        }
-    }, []);
+  useEffect(() => {
+    if (typeof window !== "undefined") {
+      const item = localStorage.getItem("token");
+      setToken(item);
+      if (!item) {
+        router.push("/login");
+      }
+    }
+  }, []);
 
-    useEffect(() => {
-        axios(`https://api-dev.blov.us/getDonorCard`, {
-            method: "GET",
-            crossDomain: true,
-            headers: {
-                Authorization: `Token ${token}`,
-            },
-        })
-            .then((res) => {
-                setCardList(res.data);
-            })
-            .catch((e) => {
-                console.log(e);
-            });
-    }, [token]);
+  useEffect(() => {
+    axios(`https://api-dev.blov.us/getDonorCard`, {
+      method: "GET",
+      crossDomain: true,
+      headers: {
+        Authorization: `Token ${token}`,
+      },
+    })
+      .then((res) => {
+        setCardList(res.data);
+      })
+      .catch((e) => {
+        console.log(e);
+      });
+  }, [token]);
+
+  useEffect(() => {
+    if (active) {
+      setTimeout(() => setActive(false), 2000);
+    }
+  }, [active]);
+  const nftRef = useRef();
+  const myNft = nftRef.current;
+  const handleDownloadNFT = () => {
+    domtoimage.toBlob(myNft).then((blob) => {
+      saveAs(blob, "my_NFT_blood_donation_from_BLOV.png");
+    });
+    setActive((active) => !active);
+  };
   return (
     <>
       <Layout>
@@ -91,7 +122,13 @@ export default function Wallet() {
             </NTFborder>
           ))}
         </Carousel>
+        <ShareLogoBox onClick={handleDownloadNFT}>
+          {" "}
+          <ShareLogo src="/mywallet/share-icon.svg" />
+        </ShareLogoBox>
+        {active && <Toast msg={"저장 완료!"} width={"100%"} />}
       </Layout>
+
       <Footer />
     </>
   );
